requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncmy==0.2.10
bcrypt==3.2.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
motor==3.7.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
pytest==8.3.4
python-dotenv==1.0.1
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
